Extract shared cup logic into CupCompetitionHandler base class
Deduplicate ~466 lines of near-identical code across the 4 cup handlers
(KnockoutCup, SwissFormat, GroupStageCup, LeagueWithPlayoff) into an
abstract base class. Shared methods: getRedirectRoute, getHybridMatchBatch,
resolveCompletedTies, createTie, roundExists, getCurrentRound, and
maybeResetYellowCards. Pure refactoring — no behavior change.
