+++ app/Modules/Match/Handlers/LeagueWithPlayoffHandler.php
@@ -2,17 +2,14 @@
 
 namespace App\Modules\Match\Handlers;
 
-use App\Modules\Competition\Contracts\CompetitionHandler;
 use App\Modules\Competition\Contracts\PlayoffGenerator;
-use App\Modules\Competition\DTOs\PlayoffRoundConfig;
 use App\Modules\Competition\Playoffs\PlayoffGeneratorFactory;
-use App\Modules\Match\Events\CupTieResolved;
 use App\Modules\Match\Services\CupTieResolver;
+use App\Modules\Squad\Services\EligibilityService;
 use App\Models\CupTie;
 use App\Models\Game;
 use App\Models\GameMatch;
 use Illuminate\Support\Collection;
-use Illuminate\Support\Str;
 
 /**
  * Handler for league competitions that have end-of-season playoffs.
@@ -20,12 +17,15 @@ use Illuminate\Support\Str;
  * Combines regular league handling with knockout playoff generation
  * and resolution after the regular season ends.
  */
-class LeagueWithPlayoffHandler implements CompetitionHandler
+class LeagueWithPlayoffHandler extends CupCompetitionHandler
 {
     public function __construct(
-        private PlayoffGeneratorFactory $playoffFactory,
-        private CupTieResolver $tieResolver,
-    ) {}
+        CupTieResolver $tieResolver,
+        EligibilityService $eligibilityService,
+        private readonly PlayoffGeneratorFactory $playoffFactory,
+    ) {
+        parent::__construct($tieResolver, $eligibilityService);
+    }
 
     public function getType(): string
     {
@@ -34,21 +34,7 @@ class LeagueWithPlayoffHandler implements CompetitionHandler
 
     public function getMatchBatch(string $gameId, GameMatch $nextMatch): Collection
     {
-        // Return all matches for the same matchday (league) or date (playoffs)
-        return GameMatch::with(['homeTeam', 'awayTeam', 'cupTie'])
-            ->where('game_id', $gameId)
-            ->where('competition_id', $nextMatch->competition_id)
-            ->where('played', false)
-            ->where(function ($query) use ($nextMatch) {
-                if ($nextMatch->cup_tie_id) {
-                    // Playoff match - get all matches on same date
-                    $query->whereDate('scheduled_date', $nextMatch->scheduled_date->toDateString());
-                } else {
-                    // League match - get all matches in same matchday
-                    $query->where('round_number', $nextMatch->round_number);
-                }
-            })
-            ->get();
+        return $this->getHybridMatchBatch($gameId, $nextMatch);
     }
 
     public function beforeMatches(Game $game, string $targetDate): void
@@ -58,33 +44,19 @@ class LeagueWithPlayoffHandler implements CompetitionHandler
             return;
         }
 
-        // Check if we should generate the next playoff round
         if ($this->shouldGeneratePlayoffRound($game, $generator)) {
-            $nextRound = $this->getCurrentPlayoffRound($game, $generator) + 1;
+            $nextRound = $this->getCurrentRound($game->id, $generator->getCompetitionId()) + 1;
             $this->generatePlayoffRound($game, $generator, $nextRound);
         }
     }
 
     public function afterMatches(Game $game, Collection $matches, Collection $allPlayers): void
     {
-        // Resolve any completed playoff ties
         $playoffMatches = $matches->filter(fn ($m) => $m->cup_tie_id !== null);
 
         if ($playoffMatches->isNotEmpty()) {
-            $this->resolvePlayoffTies($game, $playoffMatches, $allPlayers);
+            $this->resolveCompletedTies($game, $playoffMatches, $allPlayers);
         }
-    }
-
-    public function getRedirectRoute(Game $game, Collection $matches, int $matchday): string
-    {
-        $firstMatch = $matches->first();
-
-        return route('game.results', array_filter([
-            'gameId' => $game->id,
-            'competition' => $firstMatch->competition_id ?? $game->competition_id,
-            'matchday' => $firstMatch->round_number ?? $matchday,
-            'round' => $firstMatch?->round_name,
-        ]));
     }
 
     /**
@@ -92,7 +64,6 @@ class LeagueWithPlayoffHandler implements CompetitionHandler
      */
     public function isSeasonComplete(Game $game): bool
     {
-        // First check if all league matches are played
         $unplayedLeague = GameMatch::where('game_id', $game->id)
             ->where('competition_id', $game->competition_id)
             ->whereNull('cup_tie_id')
@@ -103,10 +74,9 @@ class LeagueWithPlayoffHandler implements CompetitionHandler
             return false;
         }
 
-        // Then check playoff completion
         $generator = $this->playoffFactory->forCompetition($game->competition_id);
         if (!$generator) {
-            return true; // No playoffs configured
+            return true;
         }
 
         return $generator->isComplete($game);
@@ -117,7 +87,6 @@ class LeagueWithPlayoffHandler implements CompetitionHandler
      */
     private function shouldGeneratePlayoffRound(Game $game, PlayoffGenerator $generator): bool
     {
-        // Check if regular season is complete
         $regularSeasonComplete = !GameMatch::where('game_id', $game->id)
             ->where('competition_id', $game->competition_id)
             ->whereNull('cup_tie_id')
@@ -134,135 +103,37 @@ class LeagueWithPlayoffHandler implements CompetitionHandler
             return false;
         }
 
-        $currentRound = $this->getCurrentPlayoffRound($game, $generator);
+        $competitionId = $generator->getCompetitionId();
+        $currentRound = $this->getCurrentRound($game->id, $competitionId);
         $nextRound = $currentRound + 1;
 
-        // Check if we've already generated all rounds
         if ($nextRound > $generator->getTotalRounds()) {
             return false;
         }
 
         // For round 1, generate if it doesn't exist
         if ($nextRound === 1) {
-            return !$this->playoffRoundExists($game, $generator, 1);
+            return !$this->roundExists($game->id, $competitionId, 1);
         }
 
         // For later rounds, check if previous round is complete
         $previousRoundComplete = CupTie::where('game_id', $game->id)
-            ->where('competition_id', $generator->getCompetitionId())
+            ->where('competition_id', $competitionId)
             ->where('round_number', $currentRound)
             ->where('completed', false)
             ->doesntExist();
 
-        return $previousRoundComplete && !$this->playoffRoundExists($game, $generator, $nextRound);
+        return $previousRoundComplete && !$this->roundExists($game->id, $competitionId, $nextRound);
     }
 
-    /**
-     * Generate fixtures for a playoff round.
-     */
     private function generatePlayoffRound(Game $game, PlayoffGenerator $generator, int $round): void
     {
+        $competitionId = $generator->getCompetitionId();
         $config = $generator->getRoundConfig($round, $game->season);
         $matchups = $generator->generateMatchups($game, $round);
 
         foreach ($matchups as [$homeTeamId, $awayTeamId]) {
-            $this->createPlayoffTie($game, $generator, $homeTeamId, $awayTeamId, $config);
+            $this->createTie($game, $competitionId, $homeTeamId, $awayTeamId, $config);
         }
-    }
-
-    /**
-     * Create a playoff tie with its matches.
-     */
-    private function createPlayoffTie(
-        Game $game,
-        PlayoffGenerator $generator,
-        string $homeTeamId,
-        string $awayTeamId,
-        PlayoffRoundConfig $config,
-    ): void {
-        $tie = CupTie::create([
-            'id' => Str::uuid()->toString(),
-            'game_id' => $game->id,
-            'competition_id' => $generator->getCompetitionId(),
-            'round_number' => $config->round,
-            'home_team_id' => $homeTeamId,
-            'away_team_id' => $awayTeamId,
-        ]);
-
-        // Create first leg match
-        $firstLeg = GameMatch::create([
-            'id' => Str::uuid()->toString(),
-            'game_id' => $game->id,
-            'competition_id' => $generator->getCompetitionId(),
-            'round_name' => $config->name,
-            'round_number' => $config->round,
-            'home_team_id' => $homeTeamId,
-            'away_team_id' => $awayTeamId,
-            'scheduled_date' => $config->firstLegDate,
-            'cup_tie_id' => $tie->id,
-        ]);
-
-        $tie->update(['first_leg_match_id' => $firstLeg->id]);
-
-        // Create second leg if two-legged
-        if ($config->twoLegged && $config->secondLegDate) {
-            $secondLeg = GameMatch::create([
-                'id' => Str::uuid()->toString(),
-                'game_id' => $game->id,
-                'competition_id' => $generator->getCompetitionId(),
-                'round_name' => $config->name . '_return',
-                'round_number' => $config->round,
-                'home_team_id' => $awayTeamId, // Teams swap for second leg
-                'away_team_id' => $homeTeamId,
-                'scheduled_date' => $config->secondLegDate,
-                'cup_tie_id' => $tie->id,
-            ]);
-
-            $tie->update(['second_leg_match_id' => $secondLeg->id]);
-        }
-    }
-
-    /**
-     * Resolve completed playoff ties.
-     */
-    private function resolvePlayoffTies(Game $game, Collection $playoffMatches, Collection $allPlayers): void
-    {
-        $tieIds = $playoffMatches->pluck('cup_tie_id')->unique()->filter();
-
-        foreach ($tieIds as $tieId) {
-            $tie = CupTie::with(['firstLegMatch', 'secondLegMatch', 'competition'])->find($tieId);
-
-            if (!$tie || $tie->completed) {
-                continue;
-            }
-
-            $winnerId = $this->tieResolver->resolve($tie, $allPlayers);
-
-            if ($winnerId) {
-                $match = $tie->secondLegMatch ?? $tie->firstLegMatch;
-                CupTieResolved::dispatch($tie, $winnerId, $match, $game, $tie->competition);
-            }
-        }
-    }
-
-    /**
-     * Get the current (highest) playoff round that has been generated.
-     */
-    private function getCurrentPlayoffRound(Game $game, PlayoffGenerator $generator): int
-    {
-        return CupTie::where('game_id', $game->id)
-            ->where('competition_id', $generator->getCompetitionId())
-            ->max('round_number') ?? 0;
-    }
-
-    /**
-     * Check if a playoff round already exists.
-     */
-    private function playoffRoundExists(Game $game, PlayoffGenerator $generator, int $round): bool
-    {
-        return CupTie::where('game_id', $game->id)
-            ->where('competition_id', $generator->getCompetitionId())
-            ->where('round_number', $round)
-            ->exists();
     }
 }
